+++ convertkit/admin/section/class-convertkit-admin-section-general.php
@@ -15,15 +15,6 @@
 class ConvertKit_Admin_Section_General extends ConvertKit_Admin_Section_Base {
 
 	/**
-	 * Holds the API instance.
-	 *
-	 * @since   1.9.6
-	 *
-	 * @var     ConvertKit_API_V4
-	 */
-	private $api;
-
-	/**
 	 * Holds the ConvertKit Account Name.
 	 *
 	 * @since   1.9.6
@@ -147,19 +138,10 @@ class ConvertKit_Admin_Section_General extends ConvertKit_Admin_Section_Base {
 			exit();
 		}
 
-		// Initialize the API.
-		$this->api = new ConvertKit_API_V4(
-			CONVERTKIT_OAUTH_CLIENT_ID,
-			CONVERTKIT_OAUTH_CLIENT_REDIRECT_URI,
-			$this->settings->get_access_token(),
-			$this->settings->get_refresh_token(),
-			$this->settings->debug_enabled(),
-			'settings'
-		);
-
 		// Get Account Details, which we'll use in account_name_callback(), but also lets us test
 		// whether the API credentials are valid.
-		$this->account = $this->api->get_account();
+		$account       = new ConvertKit_Resource_Account();
+		$this->account = $account->refresh();
 
 		// If the request succeeded, no need to perform further actions.
 		if ( ! is_wp_error( $this->account ) ) {
@@ -224,6 +206,7 @@ class ConvertKit_Admin_Section_General extends ConvertKit_Admin_Section_Base {
 		}
 
 		// Delete cached resources.
+		$account         = new ConvertKit_Resource_Account();
 		$creator_network = new ConvertKit_Resource_Creator_Network_Recommendations();
 		$custom_fields   = new ConvertKit_Resource_Custom_Fields();
 		$forms           = new ConvertKit_Resource_Forms();
@@ -232,6 +215,7 @@ class ConvertKit_Admin_Section_General extends ConvertKit_Admin_Section_Base {
 		$products        = new ConvertKit_Resource_Products();
 		$sequences       = new ConvertKit_Resource_Sequences();
 		$tags            = new ConvertKit_Resource_Tags();
+		$account->delete();
 		$creator_network->delete();
 		$custom_fields->delete();
 		$forms->delete();
@@ -695,8 +679,19 @@ class ConvertKit_Admin_Section_General extends ConvertKit_Admin_Section_Base {
 			return;
 		}
 
-		// Also refresh Landing Pages, Tags and Posts. Whilst not displayed in the Plugin Settings, this ensures up to date
+		// Also refresh other resources. Whilst not displayed in the Plugin Settings, this ensures up to date
 		// lists are stored for when editing e.g. Pages.
+
+		// Refresh Custom Fields.
+		$custom_fields = new ConvertKit_Resource_Custom_Fields( 'settings' );
+		$result        = $custom_fields->refresh();
+
+		// Bail if an error occured.
+		if ( is_wp_error( $result ) ) {
+			return;
+		}
+
+		// Refresh Landing Pages.
 		$landing_pages = new ConvertKit_Resource_Landing_Pages( 'settings' );
 		$result        = $landing_pages->refresh();
 
@@ -705,6 +700,7 @@ class ConvertKit_Admin_Section_General extends ConvertKit_Admin_Section_Base {
 			return;
 		}
 
+		// Refresh Posts.
 		remove_all_actions( 'convertkit_resource_refreshed_posts' );
 		$posts  = new ConvertKit_Resource_Posts( 'settings' );
 		$result = $posts->refresh();
@@ -714,6 +710,7 @@ class ConvertKit_Admin_Section_General extends ConvertKit_Admin_Section_Base {
 			return;
 		}
 
+		// Refresh Products.
 		$products = new ConvertKit_Resource_Products( 'settings' );
 		$result   = $products->refresh();
 
@@ -722,6 +719,7 @@ class ConvertKit_Admin_Section_General extends ConvertKit_Admin_Section_Base {
 			return;
 		}
 
+		// Refresh Sequences.
 		$sequences = new ConvertKit_Resource_Sequences( 'settings' );
 		$result    = $sequences->refresh();
 
@@ -730,6 +728,7 @@ class ConvertKit_Admin_Section_General extends ConvertKit_Admin_Section_Base {
 			return;
 		}
 
+		// Refresh Tags.
 		$tags   = new ConvertKit_Resource_Tags( 'settings' );
 		$result = $tags->refresh();